+++ end_to_end_test/lib/serializers.dart
@@ -50,6 +50,7 @@ part 'serializers.g.dart';
   TestEnum,
   UsesHandCoded,
   ValueUsingImportAs,
+  ValueWithCustomSerializer,
   ValueWithHasInt,
   ValueWithInt,
   WireNameEnum,